validation: pass block validator reference to individual tx validators
This way the tx validator can still use the optimizations provided by fetching all
data at once in the block validator, but it is only done when actually needed.
Without this it would always fetch the data even when it's just doing a syntax validation,
i.e. when storing blocks before the last checkpoint.