[Fix] Empty Transaction Parsing
Currently, transaction processing in the empty transaction case causes an exception `TypeError: no
implicit conversion of nil into Integer`. This error occurs because segwit enabled transaction
parsing does not correctly handle the case where the transaction has no inputs but is not a valid
segwit transaction.

This fix provides a test to reproduce the error case and corrects the behavior in the above edge
case by checking that the marker and flag values conform to those required by the segwit
standard. In the case where they do not, it backs out the previous reads enabling the transaction to
parse correctly.