Support TxOut when verifying an input sig
This adds support providing a TxOut when verifying an input signature.

This is useful in one of our use cases where we need to sign
a transaction but only have enough information to craft a TxOut, not
a full Tx. Since that provides the same information as the Tx does, this
option now supports that.

Because the duck-typing was used in a few places I extracted it out to
a method than can be used in a multiple places.