Add initial rspec structure and tests
Add a new directory `spec-rspec` that will become populated with migrated RSpec
tests. The plan is that once all tests are migrated this directory can safely
be renamed to `spec` and the existing directory can be removed.

Several fixes related to warnings in openssl and secp256k1 FFI interfaces are
included as well.