Add to_json*/from_json* Deprecation Warnings
Addresses the first half of #267

Add deprecation warnings to all `to_json*` and `from_json*` methods to indicate
that these will be removed in a future release. All deprecated methods are also
moved to be near each other so they will be easier to remove in the future.

This causes the tests to display a lot of warnings and the tests and examples
using these methods must also be removed / revised once they are gone for good.